[MB-543] Update all projects to recommended settings and fixed warnings.

+++ Airship/UI/Default/Push/Classes/Shared/UAPushSettingsAddTagViewController.m
@@ -142,7 +142,7 @@ enum TagSections {
             
             // Configure the cell...
             
-            cell.textLabel.text = [self.presetTags objectAtIndex:indexPath.row];
+            cell.textLabel.text = [self.presetTags objectAtIndex:(NSUInteger)indexPath.row];
             cell.accessoryType = UITableViewCellAccessoryNone;
             
             return cell;
